pyproject.toml
[project]
name = "utahwaterpoloassociation"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "voidfiles", email = "voidfiles@gmail.com" }
]
dependencies = [
    "pelican-tailwindcss>=0.2.0",
    "httpx>=0.27.2",
    "pydantic>=2.9.2",
    "jinja2>=3.1.4",
    "black>=24.10.0",
    "autopep8>=2.3.1",
    "markdown>=3.7",
    "pyyaml>=6.0.2",
    "honcho>=2.0.0",
    "watchdog>=5.0.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/utahwaterpoloassociation"]
